+++ commands/model.go
@@ -1,11 +1,6 @@
 package commands
 
-import (
-	"fmt"
-	"io/ioutil"
-
-	"github.com/HPI-BP2015H/go-travis/client"
-)
+import "fmt"
 
 type Branches struct {
 	Branches []Branch `json:"branches"`
@@ -91,23 +86,4 @@ func (u User) String() string {
 		return fmt.Sprintf("%s (%s)", u.Login, u.Name)
 	}
 	return u.Login
-}
-
-// CurrentUser returns the user currently logged in into Travis
-func CurrentUser(client *client.Client) (User, error) {
-	user := User{}
-	res, err := client.PerformAction("user", "current", map[string]string{})
-	if err != nil {
-		return user, fmt.Errorf("Error: Could not get current user! \n%s", err.Error())
-	}
-	if res.StatusCode > 299 {
-		body, err := ioutil.ReadAll(res.Body)
-		if err != nil {
-			return user, err
-		}
-		return user, fmt.Errorf("Unexpected HTTP status: %d\n%s\n", res.StatusCode, string(body))
-	}
-	defer res.Body.Close()
-	res.Unmarshal(&user)
-	return user, nil
 }
